day 14 completed

+++ day_014/03 elif.py
@@ -1,6 +1,6 @@
 
 
-num = int(input("Enter the value of num: "))
+num = float(input("Enter the value of num: "))
 if (num < 0):
   print("Number is negative.")
 elif (num == 0):
